- fix the project cards , make it as saperated widget in order to reuse it in the microbots - start with the ui of microbots admin page (make the projects getting , but still have something to handle with it ) - adding get projects by user id , inOrder to add it to the avatar - adding the projects card to the user avatar

+++ lib/core/widgets/projects_container/custom_circular_progress.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:vision_app/core/res/color/app_colors.dart';
 
 class CustomCircularProgress extends StatelessWidget {
-  final double percentage; // from 0.0 to 1.0
+  final double percentage; // from 0.0 to 1.0, so we pass the percentaage/100
   final double size;
   final double strokeWidth;
   final Color activeColor;
@@ -33,7 +33,7 @@ class CustomCircularProgress extends StatelessWidget {
         ),
 
         CircularProgressIndicator(
-          value: clampedPercentage,
+          value: clampedPercentage.toDouble(),
           strokeWidth: strokeWidth,
           valueColor: AlwaysStoppedAnimation<Color>(activeColor),
           backgroundColor: AppColors.transparentColor,
